Les distances et les trajets des lieux sont désormais des int

+++ src/java/ordo/data/entities/Trajet.java
@@ -25,9 +25,9 @@ public class Trajet implements Serializable {
     private Long id;
     
     @Column
-    private float duree;
+    private int duree;
     @Column
-    private float distance;
+    private int distance;
     @ManyToOne
     private Lieu depart;
     @ManyToOne
@@ -48,7 +48,7 @@ public class Trajet implements Serializable {
         return duree;
     }
 
-    public void setDuree(float duree) {
+    public void setDuree(int duree) {
         this.duree = duree;
     }
 
@@ -56,7 +56,7 @@ public class Trajet implements Serializable {
         return distance;
     }
 
-    public void setDistance(float distance) {
+    public void setDistance(int distance) {
         this.distance = distance;
     }
